backend/requirements.txt
aerich
aiosqlite
annotated-types
anyio
async-generator
async-timeout
asyncpg
cffi
click
cryptography
dictdiffer
dnspython
email-validator
fastapi
h11
idna
iniconfig
iso8601
packaging
passlib
pluggy
pycparser
pydantic
pydantic_core
PyJWT
pypika-tortoise
pytest
python-dateutil
python-decouple
python-multipart
pytz
setuptools
six
sniffio
starlette
tomlkit
tortoise-orm
typing_extensions
tzdata
uuid
uvicorn
redis
redis[hiredis]
asyncio-redis
apscheduler
websockets
broadcaster
aiohttp
asyncio